Added tabbed view for companies. Stored companies and products using Realm.

+++ app/src/main/java/com/mr/app/android/storesample/data/Company.kt
@@ -3,5 +3,5 @@ package com.mr.app.android.storesample.data
 /**
  * Created by Marioara Rus on 8/21/2018.
  */
-data class Company(var id: Int,
+data class Company(var id: Long,
                    var name: String)
