+++ app/src/main/java/com/mr/app/android/storesample/data/User.kt
@@ -3,6 +3,6 @@ package com.mr.app.android.storesample.data
 /**
  * Created by Marioara Rus on 8/21/2018.
  */
-data class User(var id: Int,
+data class User(var id: Long,
                 var email: String,
                 var name: String)
